Refactor geoseries/geodataframe structure
