+++ .pre-commit-config.yaml
@@ -13,6 +13,7 @@ repos:
         hooks:
               - id: black
                 files: python/cuspatial/.*
+                args: ["--config", "pyproject.toml"]
       - repo: https://gitlab.com/pycqa/flake8
         rev: 3.7.7
         hooks:
