+++ python/cuspatial/cuspatial/__init__.py
@@ -22,8 +22,8 @@ from .core.trajectory import (
     trajectory_bounding_boxes,
     trajectory_distances_and_speeds,
 )
-from .geometry.geoseries import GeoSeries
-from .geometry.geodataframe import GeoDataFrame
+from .core.geoseries import GeoSeries
+from .core.geodataframe import GeoDataFrame
 from .io.shapefile import read_polygon_shapefile
 from .io.geopandas import from_geopandas
