+++ python/cuspatial/cuspatial/core/_column/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) 2022, NVIDIA CORPORATION
